Fold Report into ReportRunner
